move mbf dsl into test file
